+++ gradle.properties
@@ -17,7 +17,7 @@ kotlin.native.ignoreIncorrectDependencies=true
 org.gradle.daemon=true
 org.gradle.caching=true
 org.gradle.vfs.watch=true
-org.gradle.jvmargs=-Xmx4g -XX:MaxPermSize=4g -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
+org.gradle.jvmargs=-Xmx6g -XX:MaxPermSize=6g -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
 
 # kotlin
 kotlin_version=1.5.0
@@ -43,7 +43,7 @@ jwks_rsa_version=0.9.0
 # utility
 logback_version=1.2.3
 junit_version=4.12
-dokka_version=0.9.17
+dokka_version=1.4.32
 nodejs_version=10.15.1
 npm_version=6.4.1
 slf4j_version=1.7.30
