--- gradle/dokka.gradle
@@ -1,21 +1,0 @@
-apply plugin: 'org.jetbrains.dokka'
-
-dokka {
-    outputFormat = 'html'
-    outputDirectory = "$buildDir/kdoc"
-
-    kotlinTasks {
-        defaultKotlinTasks().findAll { it.name.contains("Jvm") }
-    }
-
-    reportUndocumented = false
-}
-
-task dokkaJar(type: Jar, dependsOn: 'dokka') {
-    classifier 'kdoc'
-    from fileTree(new File(project.buildDir, 'kdoc'))
-}
-
-task dokkaJavadoc {
-    // do nothing
-}
